Split up component
